Rename the files after refactoring
