+++ lightly_studio/src/lightly_studio/core/dataset_query/__init__.py
@@ -10,7 +10,7 @@ from .order_by import (
     OrderByField,
     OrderByMetadataField,
 )
-from .sample_evaluation_expression import SampleEvaluationQuery
+from .sample_evaluation_query import SampleEvaluationQuery
 from .segmentation_mask_expression import SegmentationMaskField, SegmentationMaskQuery
 from .video_sample_field import VideoSampleField
